+++ tests/IntervalTest.php
@@ -886,14 +886,16 @@ final class IntervalTest extends TestCase
     public function test_until_with_negative_duration(): void
     {
         $interval = Interval::until(Time::at(hour: 10), Duration::fromFormat('-PT3H', DurationFormat::Iso8601));
-        self::assertSame(1, $interval->duration->sign);
+        self::assertFalse($interval->duration->isZero());
+        self::assertFalse($interval->duration->isNegative());
         self::assertTrue($interval->duration->equals(Duration::of(hours: 21)));
     }
 
     public function test_around_with_negative_duration(): void
     {
         $interval = Interval::around(Time::at(hour: 10, minute: 30), Duration::fromFormat('-PT1H', DurationFormat::Iso8601));
-        self::assertSame(1, $interval->duration->sign);
+        self::assertFalse($interval->duration->isZero());
+        self::assertFalse($interval->duration->isNegative());
         self::assertTrue($interval->duration->equals(Duration::of(hours: 23)));
         self::assertEquals($interval->start, Time::at(hour: 11));
         self::assertEquals($interval->end, Time::at(hour: 10, minute: 00));